requirements.txt
streamlit
google-generativeai
selenium
webdriver-manager
python-dotenv
streamlit-ace